fix: rename directory back to Card and fix file name casing
